Move info/error snackbars to separate file
