+++ frontend/src/views/error/NotFoundView.js
@@ -1,0 +1,12 @@
+import { Typography } from '@material-ui/core'
+import React from 'react'
+
+const NotFoundView = () => (
+  <div style={{ gridArea: 'content' }}>
+    <Typography component='h1' variant='h2' align='center' color='textPrimary'>
+      Not found
+    </Typography>
+  </div>
+)
+
+export default NotFoundView
